+++ packages/app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @firebase/app
 
+## 0.7.11
+
+### Patch Changes
+
+- Update SDK_VERSION.
+
 ## 0.7.10
 
 ### Patch Changes